% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irg.R
\name{calc_irg}
\alias{calc_irg}
\title{IRG}
\usage{
calc_irg(DT, scaled = TRUE, id = "id", year = "yr")
}
\arguments{
\item{DT}{data.table of model parameters (output from model_params).}

\item{scaled}{boolean indicating if irg should be rescaled between 0-1 within id and year. If TRUE, provide id and year. Default is TRUE.}

\item{id}{id column. default is 'id'. See details.}

\item{year}{year column name. default is 'yr'.}
}
\value{
Extended data.table 'irg' column of instantaneous rate of green up calculated for each day of the year, for each individual and year.
}
\description{
Instantaneous rate of green-up.
}
\details{
The DT argument expects a data.table of model estimated parameters for double logistic function of NDVI for each year and individual. Since it is the rate of green-up, model parameters required are only xmidS and scalS.

The scaled argument is used to optionally rescale the IRG result to 0-1, for each year and individual.

The id argument is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis. This should match the id provided to filtering functions.
}
