% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_tests.R
\name{causal_stat}
\alias{causal_stat}
\title{Granger-Causal Test Statistic}
\usage{
causal_stat(root, shoot, times, theta, alternative = "twodir")
}
\arguments{
\item{root}{A \code{double} value of the vector containing the detrended and standardized measurements of the first signal (the root expressions).}

\item{shoot}{A \code{double} value of the vector containing the detrended and standardized measurements of the second signal (the shoot expressions).}

\item{times}{A \code{double} value of the vector collecting the times (measured in the same unit) at which the measurements were taken
for both signals (\code{root} and \code{shoot}).}

\item{theta}{A \code{double} value for the starting values of the parameter vector \eqn{\theta}{theta} of the full causal model (must be of length 8).}

\item{alternative}{A \code{character} value defining the alternative hypothesis for the Granger-Causal test (with the null hypothesis always being "the signals do not Granger-cause each other").
The default value is \code{"twodir"} which tests the alternative that "both signals Granger-cause each other". The other options are \code{"rtos"} which tests the alternative
"the first signal (root) Granger-causes the second signal (shoot)" and \code{"stor"} which tests the alternative "the second signal (shoot) Granger-causes the first signal (root)".}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{stat}{The value of the Granger-Causal test statistic for the specified alternative hypothesis \code{HA}}
 \item{parameters}{A \code{double} vector containing the values of the parameters relevant for the specified alternative hypothesis \code{HA}
 (i.e. intensity of the causal impact (\eqn{\psi}{psi}) and the time of maximal impact (\eqn{\gamma}{gamma}))}
}
}
\description{
This function is used within the \code{granger_test} function to compute the Granger-Causal test statistic on the signals of interest
as well as on the simulated ones based on the \code{sim_proc} function (see Heerah et al., 2020).
}
\author{
Roberto Molinari and Stephane Guerrier
}
