% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_type}
\alias{check_type}
\title{check type}
\usage{
check_type(DT = NULL, col = NULL, type = NULL)
}
\arguments{
\item{DT}{data.table.}

\item{col}{column name.}

\item{type}{typeof type type typeof type.}
}
\description{
check type
}
