% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irg.R
\name{irg}
\alias{irg}
\title{IRG}
\usage{
irg(DT)
}
\value{
Extended data.table 'irg' column of instantaneous rate of green-up calculated for each day of the year, for each individual and year.
}
\description{
IRG
}
\examples{
# Load data.table
library(data.table)

# Read in example data
ndvi <- fread(system.file("extdata", "ndvi.csv", package = "irg"))

# Calculate IRG for each day of the year
out <- irg(ndvi)
}
\seealso{
Other irg: \code{\link{calc_irg}}
}
\concept{irg}
