% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{logit2}
\alias{logit2}
\title{TO DO}
\usage{
logit2(x)
}
\arguments{
\item{x}{A \code{double} value of the vector containing the detrended and standardized measurements of the first signal (the root expressions).}
}
\value{
TO DO
}
\description{
TO DO
}
\examples{
logit2(0.5*2)
}
\author{
Roberto Molinari and Stephane Guerrier
}
