% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_params}
\alias{model_params}
\title{Model parameters}
\usage{
model_params(DT, id = "id", year = "yr", xmidS = NULL,
  xmidA = NULL, scalS = NULL, scalA = NULL)
}
\arguments{
\item{DT}{data.table of NDVI time series}

\item{id}{id column. default is 'id'. See details.}

\item{year}{year column name. default is 'yr'.}

\item{xmidS}{"spring inflection point"}

\item{xmidA}{"fall inflection point"}

\item{scalS}{"scale parameter for spring green-up portion of the NDVI curve"}

\item{scalA}{"scale parameter for fall dry-down portion of the NDVI curve"}
}
\value{
data.table of model estimated parameters for double logistic model
}
\description{
Estimated parameters for fitting double logistic curve.
}
\details{
Default value for the year column is 'yr'. If you only have one year of data, set to NULL.

The id parameter is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis. This should match the id provided to filtering functions.

Formula and arguments \code{xmidS}, \code{xmidA}, \code{scalS}, \code{scalA} following this from Bischoff et al. (2012).

\deqn{\frac{1}{1 + \exp{\frac{xmidS - t}{scalS}}} - \frac{1}{1 + \exp{\frac{xmidA - t}{scalA}}}}
}
\references{
\url{https://www.journals.uchicago.edu/doi/abs/10.1086/667590}
}
