% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irg.R
\name{calc_irg}
\alias{calc_irg}
\title{IRG}
\usage{
calc_irg(DT, scaled = TRUE, id = "id", year = "yr")
}
\arguments{
\item{DT}{data.table of model parameters (output from model_params).}

\item{scaled}{boolean indicating if irg should be rescaled between 0-1 within id and year. If TRUE, provide id and year. Default is TRUE.}

\item{id}{id column. default is 'id'. See details.}

\item{year}{year column name. default is 'yr'.}
}
\value{
Model parameter data.table appended with 'irg' column of double logistic model of NDVI for a full year. Calculated at the daily scale with the following formula from Bischoff et al. (2012).
}
\description{
Instantaneous rate of green-up.
}
\details{
The id parameter is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis. This should match the id provided to filtering functions.
}
