% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irg.R
\name{calc_irg}
\alias{calc_irg}
\title{IRG}
\usage{
calc_irg(DT, fitted = TRUE, scaled = TRUE, id = "id", year = "yr")
}
\arguments{
\item{DT}{data.table of model parameters (output from model_params).}

\item{fitted}{boolean indicating if input DT is the result of model_ndvi.}

\item{scaled}{boolean indicating if irg should be rescaled between 0-1 within id and year. If TRUE, provide id and year. Default is TRUE.}

\item{id}{id column. default is 'id'. See details.}

\item{year}{year column name. default is 'yr'.}
}
\value{
Extended data.table 'irg' column of instantaneous rate of green-up calculated for each day of the year, for each individual and year.
}
\description{
Calculate the instantaneous rate of green-up.
}
\details{
The DT argument expects a data.table of model estimated parameters for double logistic function of NDVI for each year and individual. Since it is the rate of green-up, model parameters required are only xmidS and scalS.

The scaled argument is used to optionally rescale the IRG result to 0-1, for each year and individual.

The id argument is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis. This should match the id provided to filtering functions.
}
\examples{
# Load data.table
library(data.table)

# Read in example data
ndvi <- fread(system.file("extdata", "ndvi.csv", package = "irg"))

# Filter and scale NDVI time series
filter_ndvi(ndvi)
scale_doy(ndvi)
scale_ndvi(ndvi)

# Double logistic model parameters given starting parameters for nls

## Option 1: fitted = FALSE
# Calculate IRG for each day of the year
calc_irg(mods)

## Option 2: fitted = TRUE
mods <- model_params(
  ndvi,
  xmidS = 0.44,
  xmidA = 0.80,
  scalS = 0.05,
  scalA = 0.01
)

model_ndvi(mods)




# Calculate IRG for each day of the year
calc_irg(mods)
}
\seealso{
Other irg: \code{\link{irg}}
}
\concept{irg}
