% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_qa}
\alias{filter_qa}
\title{Filter with QA Band}
\usage{
filter_qa(DT, qa = "SummaryQA", good = c(0, 1))
}
\arguments{
\item{DT}{data.table of NDVI time series}

\item{qa}{QA column. default is 'SummaryQA'.}

\item{good}{values which correspond to quality pixels. default is 0 and 1.}
}
\description{
Using MODIS QA band information, filter the NDVI time series.
}
\examples{
filter_qa(DT, qa = 'SummaryQA', good = c(0, 1))
}
