% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{irg-package}
\alias{irg}
\alias{irg-package}
\title{irg: Granger-Causal Tests for Irregularly Sampled Signals}
\description{
This package implements the framework proposed in Heerah et al. (2020) to perform Granger-Causal tests on (short) bivariate and irregularly sampled signals (with same sampling times for each signal). The main functions allow to plot the two signals of interest and to perform a Granger-Causal test for the different directions of possible influence.
}
\details{
The data sets in this package may change at a moments notice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SMAC-Group/irg}
  \item \url{https://www.biorxiv.org/content/10.1101/2020.06.15.152819v1}
  \item Report bugs at \url{https://github.com/SMAC-Group/irg/issues}
}

}
\author{
\strong{Maintainer}: Roberto Molinari \email{roberto.molinari@hotmail.it} [copyright holder]

Authors:
\itemize{
  \item Stephane Guerrier [copyright holder]
}

}
