% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_ndvi}
\alias{model_ndvi}
\title{Model NDVI time series}
\usage{
model_ndvi(DT)
}
\arguments{
\item{DT}{data.table of model parameters (output from model_params).}
}
\description{
Fit double logistic model to NDVI time series given parameters estimated with model_params.
}
