% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_proc}
\alias{sim_proc}
\title{Simulate Bivariate Irregularly Sampled Signals}
\usage{
sim_proc(
  theta,
  times,
  burn_length = 3 * length(times),
  burn_step = median(diff(times))
)
}
\arguments{
\item{theta}{A \code{double} value of the vector containing the detrended and standardized measurements of the first signal (the root expressions).}

\item{times}{A \code{double} value of the vector containing the detrended and standardized measurements of the second signal (the shoot expressions).}

\item{burn_length}{A \code{double} value of the vector collecting the times (measured in the same unit) at which the measurements were taken
for both signals (\code{root} and \code{shoot}).}

\item{burn_step}{A \code{double} value for the starting values of the parameter vector \eqn{\theta}{theta} of the full causal model (must be of length 8).}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{root}{The value of the Granger-Causal test statistic for the specified alternative hypothesis \code{HA}}
 \item{shoot}{A \code{double} vector containing the values of the parameters relevant for the specified alternative hypothesis \code{HA}
 (i.e. intensity of the causal impact (\eqn{\psi}{psi}) and the time of maximal impact (\eqn{\gamma}{gamma}))}
}
}
\description{
TO DO
}
\details{
TO DO
}
\author{
Roberto Molinari and Stephane Guerrier
}
