% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_example_ee_script}
\alias{use_example_ee_script}
\title{Use an example sampling script for Earth Engine}
\usage{
use_example_ee_script(sensor = "MODIS", filepath = NULL, overwrite = FALSE)
}
\arguments{
\item{sensor}{either 'MODIS' or 'Landsat'}

\item{filepath}{file path relative to current working director, indicating
where to save the example script. default is NULL, simply printing lines
to the console.}

\item{overwrite}{boolean indicating if the file should overwrite existing
files. default is FALSE.}
}
\value{
\code{use_example_ee_script} prints an example NDVI extraction script or
if \code{filepath} is provided, saves it at the location specified.
}
\description{
Provides an example script for use in Earth Engine, as a preceeding step
to using the \code{irg} package. Use the script to sample NDVI in Earth
Engine, then use the \code{irg} package to calculate the instantaneous rate of green-up.
}
\examples{
library(irg)
use_example_ee_script(sensor = 'MODIS')
}
